pv/binding/device.cpp: add support for SR_CONF_CONTINUOUS
The dslogic allows the user to select which capture will be used.
If SR_CONF_CONTINUOUS is true, data is streamed to the usb interface,
If SR_CONF_CONTINUOUS is false, data is buffered in the internal ram.

adding this option here displays the option in pulseview ui if the
device supports it.

Signed-off-by: Diego Asanza <f.asanza@gmail.com>